Refactor error-checking code to separate file
